pyproject.toml
[project]
name = "xtract"
version = "0.1.0"
description = "Document entity extraction and evaluation framework"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["xtract*"]
exclude = ["tests*", "examples*", "config*", "data*", "output*"]

[tool.setuptools.package-data]
xtract = ["*.py"]